Day 14
This one was a bit trickier.  Part 1 was easy.  Part 2 was just annoying.  Having to detect the cycle and assume it exists.  I didn't try and do anything special in detecting the movement of boulders or anything, literally just ran some number of cycles and then saw that there was a cycle of a certain length so figured out where the big number was in that cycle.