Day 22
Just a lot of math for Part 1, didn't try and do anything clever while computing the values

Part 2, started with scanning each pattern against the diffs for each chain of prices.  That was too slow.  Switched to computing a map of length 4 chains to the value and stored in a dictionary.  That was much faster.  Then I still check the 19^4 different possibilities, but the actual check is then just the lookups in the table and summing values.

Test Plan: ran in 35ish seconds

Reviewers: 

Pull Request: https://github.com/dmmiller/adventofcode/pull/73